rename utils.py to conftest.py to follow pytest convention
